+++ .gitignore
@@ -3,3 +3,4 @@
 .idea/
 app/.DS_Store
 node_modules/
+dist/
