+++ README.md
@@ -44,3 +44,23 @@ npm start
 на Мобильном устройсте с ОС `Android`
 
 
+#Типизация
+
+В отчёте отразить
+Трудоёмкость перевода проекта на TypeScript. Самые сложные моменты в работе.
+Какие в процессе перевода были найдены ошибки.
+Решили ли вы вливать данный PR, или предпочитаете работать с JavaScript? Почему?
+
+>Для запуска:
+```
+npm i 
+npm run tsc
+npm start
+```
+
+* Трудоёмкость оказалась выше, чем предпологалось в начале. 
+    * Сложные моменты:
+        * Сложности с импоротом сторонних библотек 
+        * С миграцией адаптивных изообржений
+* Возможно были мелкие ошибки, которые не были мною зафиксированы.
+* Решил не вливать данный PR, из за сформировавшейся архитектуры кодовой базы проекта.
