+++ app/server.ts
@@ -1,10 +1,10 @@
-const http = require('http');
-const express = require('express');
-const fs = require('fs');
-const cors = require('cors');
-const api = require('./api');
-const app = express();
+import { Request, Response } from 'express';
+import express from 'express';
+import http from 'http';
+import cors from 'cors';
+const app: express.Application = express();
 const date = new Date();
+const api = require('./api');
 
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
@@ -12,14 +12,14 @@ app.use(express.urlencoded({ extended: true }));
 app
     .use(cors())
     .use(express.static('./'))
-    .get('/', function(req, res) {
+    .get('/', (req: Request, res: Response) => {
         res.sendFile('index.html');
     })
-    .get('/status', (request, response) => {
+    .get('/status', (request: Request, response: Response) => {
         response.send(api.getTime(date))
     })
-    .get('/api/events', (request, response) => {
-        let hasTypes = request.query.type,
+    .get('/api/events', (request: Request, response: Response) => {
+        const hasTypes = request.query.type,
             params = {
                 limit: request.query.limit,
                 offset: request.query.page,
@@ -28,20 +28,20 @@ app
 
         api.getEvents(params, response)
     })
-    .post('/api/events', (request, response) => {
-        let hasTypes = request.body.type,
+    .post('/api/events', (request: Request, response: Response) => {
+        const hasTypes = request.body.type,
             params = {
                 types: hasTypes && hasTypes.split(':')
             };
 
         api.getEvents(params, response)
     })
-    .get('*', (request, response) => {
+    .get('*', (request: Request, response: Response) => {
         response.status(404).sendFile(__dirname.concat('/routing/nopage/index.html'));
     });
 
-var server = http.createServer(app);
-server.listen(8000, (err) => {
+const server = http.createServer(app);
+server.listen(8000, (err: string) => {
     if (err) {
         return console.log('something happened', err)
     }
