+++ js/types/utils.types.ts
@@ -1,0 +1,4 @@
+export interface Pointer {
+    x: number,
+    y: number
+}
