+++ package.json
@@ -17,11 +17,25 @@
     "start": "run-p server murphy open",
     "murphy": "murphy",
     "test": "echo \"Error: no test specified\" && exit 1",
-    "server": "node app/server.js",
-    "open": "open http://localhost:8000"
+    "server": "ts-node app/server.ts",
+    "open": "open http://localhost:8000",
+    "tsc": "tsc -m amd",
+    "watch-ts": "tsc -w",
+    "format": "prettier --write \"js/**/*.ts\" \"js/**/*.js\"",
+    "lint": "tslint -p tsconfig.json"
   },
   "devDependencies": {
-    "eslint": "^5.6.1"
+    "@types/cors": "^2.8.4",
+    "@types/express": "^4.16.0",
+    "@types/hls.js": "^0.10.2",
+    "@types/node": "^10.12.0",
+    "eslint": "^5.6.1",
+    "hls.js": "^0.11.0",
+    "prettier": "^1.14.3",
+    "ts-node": "^7.0.1",
+    "tslint": "^5.11.0",
+    "tslint-config-prettier": "^1.15.0",
+    "typescript": "^3.1.3"
   },
   "repository": {
     "type": "git",
